Fix latest commit - add missing release config file.

+++ data/src/main/java/com/piotrek1543/android/boilerplate/data/model/CoordEntity.kt
@@ -5,7 +5,7 @@ package com.piotrek1543.android.boilerplate.data.model
  * Representation for a [CoordEntity] fetched from an external layer data source
  */
 data class CoordEntity(
-        val cityId: Long? = null,
+        val cityId: Int? = null,
         val lon: Double? = null,
         val lat: Double? = null
 )
